Replace werkzeug dependency with basic string checks
werkzeug is ~19k lines of Python, which is a very heavy dependency just
to parse a content-type header. The format of the header is pretty
simple, via MDN[1] the syntax roughly is:

 Content-Type: text/html; charset=utf-8
 Content-Type: multipart/form-data; boundary=something

All we want to know is whether the type is JSON or not, so we can just
check if the mime part of the value is "application/json". Originally I
implemented this as a regex (`^application/json($|;)`), but realized
that wasn't even necessary.

This code is already covered by tests, but I added a few to demonstrate
basic functionality of the new function.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type

Fixes #110.