Revert "Remove unnecessary work around segmentation faults :)"
It turns out that segmentation faults that I fixed were not the only
cause of flakiness. I haven't found any obvious reason why timeouts
would stop happening when tests run in isolation from each other,
but that seems to work. Remains the occasional segfault, that should
be the exception rather than a regular event.

Restoring the workaround until it becomes worth investigating further.

Trade-off: the functional tests are now 65% longer than the longest job
in the pipeline, which extends the overall time accordingly. The regular
failures did render the time-savings moot, so this actually making
things better, just not as good as they would be without flakiness.

This reverts commit a74c1ef71990601caf16323aea9080c0da39181f.