Have CI wait for HTTP to be up on server dev environment
The tcp port is available as soon as the container is running, but we
actually want to wait for the HTTP server to be up, which means all the
Rust code has been built and sources generated. dockerize already
supports that, we just need to use the "http" protocol instead of "tcp".

This allows us to get rid of the manual "sleep 30" as well.

Fixes #183.