WIP: Migrate dependency management to poetry
Switch dependency management to use poetry, which is much nicer than
pip-tools. This is largely based off of
<https://github.com/freedomofpress/securedrop-proxy/pull/122> and
applies the same changes to the Makefile and CI.

The main difference is that we install PyQt5 via system packages, so the
two versions we specifically need are listed in pyproject.toml and vary
on the Python versions shipped with bullseye and bookworm.

We also don't actually need the pathlib2 dependency any more because
our Python version is high enough, so it's dropped now that poetry
is smart enough to detect it!