Rewrite proxy in Rust, with v2 behavior
We identified multiple features that the current implementation of the
proxy is unable to support (e.g. progress reporting, resuming
downloads), necessitating a new protocol dubbed "v2". We are taking the
opportunity to rewrite this component in Rust.

Normal operation is mostly the same, input is received as a JSON blob
(parsed and validated by serde) and output in our custom JSON response
format. The `url` crate (from servo) is used to assemble and validate
the target URL, and `reqwest` fires off the HTTP request.

Downloads are now streamed back to the client over stdout, and metadata
passed over stderr.

Tests and further integration will happen in follow-up commits.

Refs #1678.

Co-authored-by: Kunal Mehta <legoktm@debian.org>