Handle `udisksctl mount` not returning excepted output
mypy noted a theoretical case in which `udisksctl mount` exits with a 0
status code but doesn't return our expected output. In this case just
raise an exception as we couldn't figure out the mount point.

While we're at it, add a test case to verify the mountpoint is parsed
correctly.