Have CI run piuparts
piuparts tests package installation, upgrade and removal. We don't care
that much about the last one so it's just a warning for now.

For each package, piuparts will install the package CI just built and
ensure that goes smoothly, making sure all the dependencies are present
and the postinst works. Issues that were found are being fixed in the
next commit.

At a later stage, once the released packages are passing piuparts too,
we can have it automatically grab the latest package from
apt.freedom.press, install that, and then test the upgrade path to the
new CI-built packages.

This CI setup relies on running docker-in-docker, which is hacky but
seems to work. Unfortunately it doesn't work for the client (starts
apparmor), workstation-config (starts systemd) and workstation-viewer
(pulls in grsec kernel). In the future we can look at switching into a
chrooted setup that might fix the last two.

Getting this to work in GitHub Actions was inspired by
<https://github.com/evgeni/action-piuparts/> but I didn't really copy
any code.

Refs #1785.