Fix issues caught by piuparts
Both -client and -export were missing dependencies on
desktop-file-utils, which is needed for the `update-desktop-database`
call in the postinst. -export also needed shared-mime-info.

As far as I can tell, the -dev and libtool-bin dependencies in -export
are unnecessary; we aren't compiling any C/C++ code at runtime that
should need them.

Finally, make the postrm commands optional per
<https://www.debian.org/doc/debian-policy/ch-relationships.html#binary-dependencies-depends-recommends-suggests-enhances-pre-depends>,
which states that "There is no guarantee that package dependencies will
be available when postrm is run".