Centralize __version__ management
There is now a single `update_version.sh` script that updates the
`__version__` variables in the client and export components and
increments the debian/changelog file.

The proxy version is hardcoded since it's going away as part of proxy v2
(also the usage in the response object was dubious to begin with).

Fixes #1815.