feat(Config): read from QubesDB if available; otherwise from environment variables
This has the nice side effect of simplifying tests: we have to care only
about getting the values we expect, not about whether certain files are
in the right place, have the right contents, &c.

For parity with the Rust implementation, we error out on zero-length values
from QubesDB.

Config is now a dataclass, which lets us enforce typing inside the Config
implementation and eliminates the need to annotate it "Any" where instantiated.

Also remove the unused Config.is_valid() and just test Config.load().