Read proxy origin from QubesDB
We are setting the proxy's origin in QubesDB (via dom0 salt), and as
discussed/agreed upon in
<https://github.com/freedomofpress/securedrop-workstation/issues/1013>,
we're reading it directly in the proxy.

We link directly to libqubesdb, generating the Rust wrapper with bindgen
so it stays in sync (at build time at least). As a result, some unsafe
is needed, but it is clearly annotated with relevant SAFETY notes.

If we're not building with QubesDB (for development purposes), we'll
continue to read from the environment and skip the bindgen and linking
steps entirely.

Fixes <https://github.com/freedomofpress/securedrop-workstation/issues/853>.

Co-authored-by: Kunal Mehta <legoktm@debian.org>