Fix lintian errors on nightlies
Because of the date-based version, we end up triggering the
"package-has-long-file-name" check, which is only an issue for
distributing them via CDs, so we can suppress them.

While we're at it, set fail-fast: false, so even if one distro fails we
can still see if the other would have succeeded or has other errors.