Remove __pycache__ folders from packages
We already delete the individual .pyc files because they're not
reproducible, but leave behind empty `__pycache__` folders. This can be
annoying if people are doing builds out of a development checkout that
has these folders that won't be present on a pristine build.

So remove the empty folders right after deleting the .pyc files.

Fixes #1907.