Provision deb822-style apt sources list
apt now supports structured deb882-style `.sources` files that can also
contain the signing key, making distribution even easier. A bit of
substitution at build time is used to inject the correct distro version
that the package is being built for.

We can also eliminate the need for removing the GPG key from trusted.gpg
before installing by doing this during the provisioning step too.

Because the key is contained in the file and therefore harder to audit
independently, a test verifies the fingerprint to ensure it's the
correct key.

There is a small downside related to testing that may need a bit more
work: the apt prod repository will now always be configured, so you have
to duplicate the file for apt-qa. It should be safe to layer packages
and repositories like this though, because versions on apt-test and
apt-qa should always have higher version numbers than prod.

Refs <https://github.com/freedomofpress/securedrop/issues/6447>.
Refs
<https://github.com/freedomofpress/securedrop-workstation/issues/876>.
Refs <https://github.com/freedomofpress/securedrop-client/issues/1788>.