Push "nightly" after each merged PR
Instead of needing to wait up to 24 hours for nightly packages
to be built, just do it instantly. This will allow for faster
development cycles because securedrop-apt-test will now reflect
what is in "main", making securedrop-workstation CI more useful too.

The only thing we need to add after the push trigger is a concurrency
limiter so multiple jobs aren't pulling/pushing to the build-logs and
securedrop-apt-test repositories at the same time.

Fixes #1959.