Fix AppArmor profile
This reverts 35e6ec8dc, since there is no "c" flag. Instead, we need to set
rw permissions on the `__pycache__` directory itself.

Also fix the filenames of the C extensions for the 3.11 switch.

For some reason alembic tries to read the OpenSSL configuration, so allow
that as well since it's harmless.

Fixes #2009.