Auto-resume failed downloads
A new _streaming_download function runs a loop that retries failed downloads,
using a configurable value from QubesDB/environment (default 3). The files
are streamed to disk in a temporary file and then loaded in memory (a FIXME
exists for that) to return a StreamedResponse with the contents as bytes.

The test works by mocking subprocess.Popen. The first time it uses Popen to
download a file, it returns 200 bytes and then raises
subprocess.TimeoutExpired, and then subsequent times it uses the real,
non-stubbed Popen with range requests.

test_download_submission_autoresume_fail makes sure that in the event of a
timeout that doesn't auto-resume, a RequestTimeoutError is ultimately thrown.

We also generate bigger random files for tests using LOADDATA_ARGS to
effectively test this.

`__reduce__` methods were added to the custom exceptions because otherwise an
exception was thrown when VCR tried to deepcopy them.