Use curl instead of dockerize in SDK CI
We just need a tool to wait until the HTTP service is working,
and it turns out we can use curl for that by running it in a
one-line bash loop, avoiding the need to download and extract
dockerize.

This approach was inspired by the securedrop.org repository CI.[1]

[1] https://github.com/freedomofpress/securedrop.org/blob/6df3cd790a8cdc7409585c30813d01a5f6a5c37c/.circleci/config.yml#L59

Fixes #1999.