Re-instate Python package discovery in securedrop-log's setup.py
This was a bad collision between 588988a ("Remove unused
securedrop-log code") and 90424f8 ("log: reorganize code to split
server from client").

The first commit removed all the Python packages from the
securedrop-log component, removing the discovery of packages was
needed. The second commit created a new Python package with the
log_server code, and relied on the line being present.

This reverts the removal in 588988a.