Specify `download_retry_limit` default in Config
Currently QubesDB lookups are required, while environment lookups are
optional. For fields with a default value, the QubesDB lookup is no
longer required.

Ideally we'd also make the environment lookups mandatory if the field is
required (`journalist_key_fingerprint`), but this causes a number of
issues with tests and should be done separately.

We were relying on `gpg_domain` being optional in env but not in
QubesDB; it's now explicitly optional at the type level.

Fixes #2041.