Don't let Range header persist into other requests
When we introduced retries with `Range` support, it was the first time
we started mutating the request headers. At the time we didn't realize
that headers are stored in a single dict that is used across requests,
so the `Range` header would get reused in other requests, and then of
course fail.

This affected any session in which a Range request was needed, so if no
download ever failed, you'd never hit it.

To fix it, we now generate a fresh dictionary on each request. This will
add a small amount of performance overhead, but it's not going to be
noticeable and more importantly, makes it impossible to re-use the
header dictionary across requests.

Add some more debug entries for future operations as well.

Fixes #2194.

(cherry picked from commit ba01651dc7b553eb8abb2c0eb46ad83fd1e5f6ed)