Have BaseErrors trigger DownloadDecryptionException in download flow
DownloadDecryptionException is specially handled in
Controller.on_file_download_failure(), so if we raise a different
exception class, the download animation will never be stopped because
file_missing is never emitted.

It is worth a proper re-examination of our exception heirarchy and
reducing the amount of indirection in the download flow, but that can be
done outside of a hotfix.

More debug logging is added to trace when slots are triggered.

Refs #1633.