sdk: Configure custom proxy VM via QubesDB/env; update docs
Instead of using a custom `/etc/sd-sdk.conf` for setting a custom proxy
VM, use our now established pattern of using QubesDB or the environment.

I am somewhat dubious of the use case of using a different proxy VM, but
I can see how it could be useful for some debugging operations.

Update and remove a lot of documentation around updating VCR cassettes
that was outdated but referenced this configuration mechanism.

Fixes #2040.