Automatically log into the dev client
Most of the time we're developing the client, we are not testing the
login screen, and just want to quickly log in. With a combination of
xdotool and oathtool, we can script typing the correct values and login.

It's a bit fragile for two reasons. First, we're just blindly typing, so
if some other window takes focus or switch, it'll be typed into that
window. Second, we're kicking off a background process that sleeps,
launching the window in the foreground, and then the background process
will do the typing into the foreground.

Also this doesn't work on Wayland; there are some alternatives that use
uinput but they have other restrictions/consequences.

It would've been easier to just have this be in the client Python code
itself, but that would've meant it also gets shipped to production,
which I think we can avoid.