Switch to builtin venv instead of virtualenv
For our purposes, venv should be a drop-in replacement for
virtualenv, with the advantage that it's maintained in the
standard library and is a bit slimmer. Diffoscoping shows
internal changes to the activation scripts and entry points
but nothing that should functionally affect us.

SecureDrop server is already using venv for its packages, so this
brings us into alignment.

Also do a little cleanup on the dependencies because lintian started
complaining a bit more; ensure that both securedrop-client and
securedrop-export have explicit `python3` dependencies, and remove
the now obsolete `${python3:Depends}` magic dependencies.