test(test_delete_sources): parameterized functional test for deletion of one or more sources
This is a complete rewrite of this test, which now (a) distinguishes
between single- and multi-select behavior and (b) checks that the
individual sources selected for deletion are confirmed and deleted, not
only the *number* of sources selected.

This includes a workaround for #2273.  Here we also change
securedrop_client.gui.widgets.BatchActionToolbar.on_action_triggered()
to call QDialog.open(), which is modal with QDialog.setModal(), rather
than QDialog.exec(), which is modal *and* blocking, which blocks pytest
timers as well as the main event loop.