Don't assume all download failures are DownloadExceptions
Exceptions can be other types, so we should use a wider type and
isinstance checks. This brings `on_message_download_failure` and
`on_reply_download_failure` in line with `on_file_download_failure`.

This is mostly a bandage over the regression seen in #2274 and doesn't
really substitute for the lack of more coordinated error handling.

Fixes #2274.