Handle get_db_object() failures as DownloadException
Our error handling expects that failures from DownloadJobs raise
DownloadExceptions, so let's do that.

This is roughly the same concept as a9f0590f746a, just in a different
part of the code. The exception handling has to exist in each class
(instead of `one_or_none()`) because we need to know the corresponding
type to pass to DownloadException.

Fixes #2274.