WIP: Display download progress for file downloads
Display a vanilla progress bar for file downloads that simply shows how
much of the file has been downloaded so far.

A new FileDownloadProgressBar widget replaces the existing animated
spinner, and we inject a signal through to the SDK to pass the current
progress back to the widget.

The widget both displays the overall total progress as a percentage and
also calculates the download speed by using an exponential moving
average to smooth it out. A timer runs every 100ms to recalculate the
speed.

A new utils.humanize_speed() is used to translate the raw bytes/second
into a human-readable version with a focus on keeping the length of the
string roughly consistent so there's less visual shifting.

TODO:
* Visual alignment issue; need padding on the right side
* tests?

Fixes #1104.