Initial support for IPP / driverless printing
Changes export logic (in sd-devices) to be able to detect IPP
printers without breaking compatibility with the old system.

To achive this, ipp-usb is used in combination with Avahi.
The former detects IPP-compatible printers and creates a local IPP
server. Avahi allows for the discovery of these printing servers
such that print dialogs can display print information.

Contrary to the legacy printer support, for driverless printing,
no print queue is setup with `lpadmin`. Printers are automatically
discovered.