Enable Debian trixie in CI; update PyQt5 dependencies
Nearly all of our CI is already parameterized by Debian version, so
adding trixie just requires adding to a bunch of lists.

The pinned PyQt5 packages in pyproject.toml are updated to reflect the
versions in trixie right now; it is expected that these will need to be
adjusted as we get closer to the actual release to stay in sync.

Other development dependencies like lxml, multidict and yarl needed to
be upgraded to versions that provide Python 3.13 wheels.