log: Don't run no-op tests
Python 3.13 now (appropriately IMO) fails if the unittest module finds
no tests. Instead of having it search for tests that don't exist, just
output a message that there are no tests.

Really removing `make test` would be the correct thing to do, but let's
leave it in place with the hope that someone will write tests soon.