Detect IPP-USB printers (via socket presence)
/var/ipp-usb/ctrl is a socket created when an IPP-USB printer
is available on the system [1]. It is internally used when running
"sudo ipp-usb check" and it can be used as a heuristic to detect
that printers are present. According to ipp-usb's man page [2]:

|  /var/ipp-usb/ctrl: ipp-usb control socket. Currently only used
|  to obtain the per-device status (printed by ipp-usb status),
|  but its functionality may be extended in a future

It also has some with some advantages over other ways to check:
  - trivial implementation (checking if a file exists is trivial)
  - no need to call subprocess
  - no need to parse output
  - very fast
  - reliable (*)

(*) The only drawback seems to be that under certain circumstances
[1] (e.g. ipp-usb crashing) the socket file can remain, even though
it can no longer detect that printers are present. However, because
ipp-usb is udev-activated, unplugging and plugging the printer's
USB does solve the issue. This is likely what most users /
administrators will do as an initial troubleshooting steps. However,
ipp-usb is not expected to crash often.

For context, a couple of alternative commands were considered:

  sudo ipp-usb check
    - Uses sudo, which is ideal to avoid
    - Requires output parsing (exit codes alone can't tell if it
      did not find printers)

  driverless list
    - Slow: (~2s) for a preflight-check / smoke test;
    - Unreliable: sometimes detects the printer and other times not;

  sudo lpinfo -v
    - Uses sudo
    - slow (~3s)
    - also detects USB printers (which could be trouble if this
      was backwards compatible with legacy printers)

  ippfind
     - slow (~1.5s)

Co-autored-by: Kunal Mehta <legoktm@debian.org>

[1]: https://github.com/OpenPrinting/ipp-usb/discussions/111
[2]: https://man.archlinux.org/man/extra/ipp-usb/ipp-usb.8.en