Fix progress bars when entire conversation is downloaded
The "Download All" menu option (or Ctrl+D shortcut) bypassed the logic
we had added in the FileWidget for connecting the progress bar to the
download job.

To fix that, we now inject the ConversationView all the way into the
DownloadConversation action, so it can iterate over all the
non-downloaded FileWidgets and begin their download.

This will hopefully be simpler in the longer-term since we can get rid
of Controller.download_conversation() and only have a single entrypoint
for downloading files: Controller.on_submission_download().

Fixes #2422.