Make progress bars fit in narrow windows
If you make the client window as narrow as it can go, when a progress
bar is displayed, the "Downloading" text gets cutoff, because there
isn't enough space for all the different elements to be displayed.

Make it fit by hiding the "Encrypted file on server" text during the
download. If the download fails, it'll get restored by the existing
`_set_file_state()`.

Fixes #2423.