Lazy-load DownloadConversation enabled state
Instead of tracking any time a file's status is updated and dynamically
updating everything in the background, we can just check whenever the
menu is opened. Whenever it's not open, we keep the action enabled so
the keyboard shortcut will work.

There appears to be no noticable performance delay in opening the
menu with the additional hook.

This gets rid of all uses of
State.selected_conversation_has_downloadable_files() except one, which
can be fixed in a follow-up using the same technique of injecting
ConversationView.