Overcome "Page range" issues on GTK
Two bugs affect GTK regarding the selection of page ranges (selecting
pages '3,4-6', for example). These bugs are:
  - PrintUnixDialog page ranges does not propagate to CUPS [1]
  - Page range when printing to IPP Printer ignores other ranges [2]

To 'solve it', the user is warned when more than 1 page range has
been selected, informing them that due to a limitation, only one
page range is supported. The document may need to be printed multiple
times if this is desired.

[1]: https://gitlab.gnome.org/GNOME/gtk/-/issues/7528
[2]: https://gitlab.gnome.org/GNOME/gtk/-/issues/7527