Use `Breaks` to remove old print-related dependencies
Now that we've switched over to IPP and the GTK print dialog, we no
longer need the old printer drivers and XPP. There's no real harm in
having them installed in the template, but it introduces a divergence
with new installs, and in the case of XPP, keeps around very
legacy/obsolete code.

We can abuse the `Breaks`[1] option to remove these packages, the Qubes
effectively runs `apt-get dist-upgrade`, so the packages will get
removed when upgrading securedrop-export.

[1] https://www.debian.org/doc/debian-policy/ch-relationships.html#packages-which-break-other-packages-breaks

Fixes #2425.