Update poetry.lock: bump setuptools to silence dependabot alert
A dependabot alert had been complaining about an outdated setuptools
dependency [1] (CVE-2025-47273).

However, it failed to bump this directly due to a poetry version
mismatch currently being fixed [2].

This PR suppresses the issue manually. poetry.lock was updated
directly, since setuptools is not a direct pyproject.toml
dependency. The command "poetry lock --no-update" did not seem
to be available on poetry 2.3.1, so this was done manually by:
  1. copying similar setuptools bump from "export/" [3]
  2. run "poetry lock" to ensure other metadata is consistent

[1]: (internal link) https://github.com/freedomofpress/securedrop-client/security/dependabot/153
[2]: https://github.com/freedomofpress/securedrop-client/pull/2461
[3]: https://github.com/freedomofpress/securedrop-client/commit/c20039ba569b546895e2962e6a6aacb65507e7a3