Connect via socks to arti in proxy mode
Assuming running in a qube connected to sys-firewall (will be a
workstation PR of its own) and with arti running in proxy mode, this
allows the proxy to connect direct to the onion address.

Needs to be couple with workstation PR that:
  - sets sd-proxy to connect to sys-firewall instead of sys-vpn
  - installs arti package in sd-proxy's template
  - starts arti in proxy mode in sd-proxy on startup
  - sets via qubesdb vm-config.SD_PROXY_ORIGIN value

NOTE: SD_PROXY_OIRIGIN must be unauthenticated (as of this commit), at a
later commit we'll add authenticated onion support with the necessary
arti changes.

Towards https://github.com/freedomofpress/securedrop-workstation/issues/456