Connect via SOCKS to Arti in proxy mode
Assuming running in a qube connected to sys-firewall (will be a
workstation PR of its own) and with arti running in proxy mode, this
allows the proxy to connect direct to the onion address.

This is coupled with workstation PR that:
  - sets sd-proxy to connect to sys-firewall instead of sys-vpn
  - installs arti package in sd-proxy's template
  - starts arti in proxy mode in sd-proxy on startup
  - sets via qubesdb vm-config.SD_PROXY_ORIGIN value

Setting the onion service authentication key is done in a follow-up.

Developers can set a `DISABLE_TOR` environment variable to
avoid using a SOCKS proxy, and both the client and app do so.

Towards https://github.com/freedomofpress/securedrop-workstation/issues/456