Install React and Redux developer tools
These are shipped as Chromium extensions, the
`electron-devtools-installer` (dev only) package takes of all the
downloading and installation process for us.

There's one bug in which the React devtools don't actually load until
you reload the page because of an upstream Electron bug
(https://github.com/electron/electron/issues/41613) but that's trivial
enough to workaround manually for now if the tab is missing.