Fix using better-sqlite3 in tests
better-sqlite3 is a native nodejs module, which means that we need it in
two versions: 1) linked against normal nodejs, and 2) linked against
Electron's nodejs.

Enable the build script in pnpm so we can download the normal nodejs
version with `pnpm rebuild better-sqlite3`.

In CI, we want to use the normal nodejs version first, so we skip the
electron-builder postinstall script first, manually rebuild (fetch) the
better-sqlite3 binary; run our tests. Then we rebuild for Electron
nodejs and do the package/build step.