Memoize most React components
This prevents them from needing to be re-rendered if the input props
don't change. Together this is sufficient to make e.g. starring a source
not require re-rendering all the unaffected sources. It does not yet fix
the overall performance of loading new source rows.

In general:
* `memo`: applied to entire components
* `useMemo`: applied to calculated values
* `useCallback`: applied to functions

The more complicated places that these wrap are good candidates for
splitting out into separate components, which'll also make profiling
them easier. But I'll defer that for another PR.

Much of this was driven by Claude Code.

Refs #2616.