Directly fetch prebuilt better-sqlite3 binaries
We need two different better-sqlite3 binaries, one for electron and one
for normal nodejs (for testing). We also need to store both so we can
switch on the fly and not need to redownload/rebuild when you want to
switch contexts.

This implements a postinstall Python script (mostly drafted by Claude),
which downloads the two binaries, and then logic in the `Database` class
to load the correct binary path basead on the runtime.

This kind of works in the direction of #2506 in that it makes the
download process way more explicit, but not complete yet since we don't
have checksum verification.

dynamic ABI