Set up systemd and packaging for Arti
The previous qubesdb-templated file was replaced with a boilerplate
python script, since a templated file is not good enough for the key
conversions necessary to convert an ctor client key to an arti one.

A securedrop-arti systemd service was added, it's based on upstream's [1],
but systemd mitigations regarding read/read-write directories were not
included due to not working out-of-the box, combined with the fact
that the proxy is expected not to keep any state.

It runs as a separate `_arti` user. For now we're not protecting it with
apparmor, but can take advantage of whatever upstream does later.

[1]: https://gitlab.torproject.org/tpo/core/arti/-/blob/084c2dee/debian/systemd/arti.service