Obtain JI onion service client credentials in proxy
Onion service key passed as "SD_PROXY_ORIGIN_KEY" via qubesdb
(production) or environment variables (testing), following project
standards.

The "get_env()" function was duplicated from securedrop-qubesdb-tools,
since it was not usable directly as a library. On top of that, its only
use-case was securedrop-whonix-config, which is now going to be
removed.

We need to convert the key from the CTor format to Arti, which isn't yet
available from upstream. This is based on @lsd-cat's initial conversion
script[1]. For more context see [2].

Fixes https://github.com/freedomofpress/securedrop-workstation/issues/1382.

[1]: https://gist.github.com/lsd-cat/d3bd05412d83628be7145462081462f9
[2]: https://github.com/freedomofpress/securedrop-workstation/issues/1382#issuecomment-3150496220

Co-authored-by: lsd-cat <giulio@gmx.com>