Replace Arti with CTor
Arti has some performance issues when connecting to onion services, so
swap in CTor for now.

The main changes are how we configure the hidden service authentication
and the systemd units.

configure_onion_service.py is now much simpler in that it doesn't
actually need to convert the private key, it just strips the prefix off
the hostname and writes it to the file.

Because we're no longer using our own securedrop-arti unit, we need to
add in an override to the tor@default unit, to be conditional on our
qubes service and dependency.

The arti packaging will be removed in a follow-up as it doesn't need to
be backported.

Co-authored-by: Kunal Mehta <legoktm@debian.org>
(cherry picked from commit 6561284daeb16ad011f7d00a5b26257764d86a41)