Capture app screenshots for translators
lib.js and extractor.js (by Claude) provide a framework that captures
all the messages that can be found, drawing a red border around them.

main.js (manually written) cycles through all the different screens,
clicking buttons and hovering over things as needed to reveal all
the different messages. Under the hood it uses playwright to drive Electron.

My ideal end goal is that in the future we get close to 100% coverage
with the screenshots and then have CI automatically upload new screenshots
to Weblate as part of continuous localization.