Add fixed data configuration for integration tests with server
With https://github.com/freedomofpress/securedrop/pull/7669, this will
allow us to start a server using a fixed dataset, in which all the data
the app sees is effectively the same each time.

The dataset does include binary encrypted files, which is unfortunately
necessary to commit because otherwise the ciphertext is intentionally
randomized and will cause subtle changes in metadata (e.g. size).

We can now have tests that make assertions based on specific UUIDs
or journalist codenames, etc., knowing that regardless of who or where
the test is running, it's effectively against the same database.