Add basic integration test that runs against a server
Login runs during globalSetup, so we can do it once and apply to all
tests. A basic test uses that login token to invoke `/api/v1/user`, which
returns data on the current user.

The test verifies both the username and expected UUID are correct,
which verifies it's running against a server using the correct fixed
data with our UUID and not a randomly generated one.

Some of this structure was copied from integration_tests/, but it's
sense diverged. The vitest config file compiles the Rust proxy, which
matches roughly how Electron does it so that we can set the correct
vite environment variables for the proxy to work.