Add basic integration test that runs against a server
Login runs as part of setup(), and then a test uses that login token
to invoke `/api/v1/user`, which returns data on the current user.

The test verifies both the username and expected UUID are correct,
which verifies it's running against a server using the correct fixed
data with our UUID and not a randomly generated one.

A lot of this structure was copied from integration_tests/.