Set stricter CSP with style nonces
Ant does inject inline styles, but we can pass a nonce to it so we're
able to ban all other inline styles. Because the header is now dynamic,
we send it as a header, which has the bonus of applying to all HTML
files in case we ever add more.

The policy itself is stricter, as we start with `default-src 'none'`
instead of 'self'. We no longer allow inline styles except those from
Ant.

In dev mode, vite's auto reload does inject inline styles and scripts,
so set a nonce for those an pass it through to the CSP header. We also
disable inlining of small assets using data: URIs so we can remove that
as well.