Have eslint forbid usage of inline styles
This is slightly too broad because e.g. setting a style parameter
on an Ant component is okay because it has access to the nonce but
in general will push people to using tailwind.

And it's just nicer to get a lint error instead of having to
figure out from CSP errors in the console log.