Support autologin in app for developers
Port over the client's `--login` which automatically enters the test/dev
credentials so you don't need to spend a lot of time (re)typing them
over and over again.

Special care is taken to only make this functionality available in dev
builds; it's guarded by `__DEV_AUTO_LOGIN__`, which is a constant that
vite knows is false in prod builds and will compile it out. This can be
verified by running, e.g.:

    grep "correct horse battery staple" out/renderer/assets/index-*.js

and finding nothing.

One slightly odd consequence of this setup is that if you explicitly
sign out, you'll log back in automatically; presumably in that case
avoid using the autologin behavior.